[MB-1031] Fix logging action situation warnings

+++ AirshipLib/AirshipLogicTests/UAPushTest.m
@@ -1569,7 +1569,7 @@ NSString *validDeviceToken = @"0123456789abcdef0123456789abcdef0123456789abcdef0
         [[self.mockedAnalytics expect] handleNotification:self.notification inApplicationState:applicationState];
         [self.push appReceivedRemoteNotification:self.notification applicationState:applicationState];
 
-        XCTAssertNoThrow([self.mockActionRunner verify], @"handleNotification should run push actions with situation %ld", expectedSituation);
+        XCTAssertNoThrow([self.mockActionRunner verify], @"handleNotification should run push actions with situation %ld", (long)expectedSituation);
         XCTAssertNoThrow([self.mockedAnalytics verify], @"analytics should be notified of the incoming notification");
 
         // Test handleNotification:fetchCompletionHandler: for every background fetch result
@@ -1591,7 +1591,7 @@ NSString *validDeviceToken = @"0123456789abcdef0123456789abcdef0123456789abcdef0
             }];
 
             XCTAssertTrue(completionHandlerCalled, @"handleNotification should call fetch completion handler");
-            XCTAssertNoThrow([self.mockActionRunner verify], @"handleNotification should run push actions with situation %ld", expectedSituation);
+            XCTAssertNoThrow([self.mockActionRunner verify], @"handleNotification should run push actions with situation %ld", (long)expectedSituation);
             XCTAssertNoThrow([self.mockedAnalytics verify], @"analytics should be notified of the incoming notification");
         }
     }
